Use PropTypes.oneOfType check instead of custom validator
We should be able to simplify this validator by using PropTypes.oneOfType.

This also will have the side effect of making the auto generated docs show the
correct PropType requirements, as it currently can't interpret our custom
validator code.

Currently we can't test this as the FixedDataTableColumn.react doesn't check
propTypes, i will look into a way to solve this.

This was identified via issue #22