Refactor and Fix scrolling (#424) (Fixes #399)
* Fix controlled scrolling not setting up the scrolling prop correctly (#399)
Earlier all it did was call the scroll handlers correctly. We also need to set/unset the scroll state. This is now done at the reducer level.

* Clean up scrolling
A lot of scroll handlers exist in the code and they dispatch multiple actions. This leads to multiple renders on the table.
Controlled scrolling was handled separately and this also made it hard to read through the code. Now they are cleaned up.